third_party/WebKit/LayoutTests/platform/win/http/tests/security/contentSecurityPolicy/filesystem-urls-match-self-expected.txt

CONSOLE ERROR: line 25: Refused to load the script 'filesystem:http://127.0.0.1:8000/temporary/fail.js' because it violates the following Content Security Policy directive: "script-src 'unsafe-inline' 'self'".

This is a testharness.js-based test.
PASS blob: does not match 'self' (see step 2 of http://www.w3.org/TR/CSP2/#match-source-expression) 
Harness: the test ran to completion.
